Create tmp folder if not present
All Apps may not have a tmp folder

+++ railties/lib/rails/tasks/dev.rake
@@ -1,14 +1,18 @@
 namespace :dev do
   desc 'Toggle development mode caching on/off'
   task :cache do
-    if File.exist? 'tmp/caching-dev.txt'
-      File.delete 'tmp/caching-dev.txt'
-      puts 'Development mode is no longer being cached.'
+    if Dir.exist? 'tmp'
+      if File.exist? 'tmp/caching-dev.txt'
+        File.delete 'tmp/caching-dev.txt'
+        puts 'Development mode is no longer being cached.'
+      else
+        FileUtils.touch 'tmp/caching-dev.txt'
+        puts 'Development mode is now being cached.'
+      end
     else
-      FileUtils.touch 'tmp/caching-dev.txt'
-      puts 'Development mode is now being cached.'
+      FileUtils.mkdir_p('tmp')  
     end
-
+    
     FileUtils.touch 'tmp/restart.txt'
   end
 end
